requirements.txt
fastapi~=0.111.0
SQLAlchemy~=2.0.30
pydantic~=2.7.1
PyJWT~=2.8.0
passlib~=1.7.4